file_system provider refactorings
