+++ internal/rules/provider/filesystem/provider_registrar.go
@@ -1,0 +1,45 @@
+package filesystem
+
+import (
+	"context"
+
+	"github.com/rs/zerolog"
+	"go.uber.org/fx"
+
+	"github.com/dadrus/heimdall/internal/config"
+	"github.com/dadrus/heimdall/internal/rules/event"
+)
+
+type registrationArguments struct {
+	fx.In
+
+	Lifecycle fx.Lifecycle
+	Config    config.Configuration
+	Queue     event.RuleSetChangedEventQueue
+}
+
+func registerProvider(args registrationArguments, logger zerolog.Logger) error {
+	if args.Config.Rules.Providers.FileSystem == nil {
+		return nil
+	}
+
+	provider, err := newProvider(args.Config.Rules.Providers.FileSystem, args.Queue, logger)
+	if err != nil {
+		return err
+	}
+
+	args.Lifecycle.Append(
+		fx.Hook{
+			OnStart: func(ctx context.Context) error {
+				return provider.Start()
+			},
+			OnStop: func(ctx context.Context) error {
+				return provider.Stop()
+			},
+		},
+	)
+
+	logger.Info().Str("_rule_provider_type", "file_system").Msg("Rule provider configured.")
+
+	return nil
+}
