+++ internal/rules/provider/filesystem/provider_test.go
@@ -5,140 +5,16 @@ import (
 	"testing"
 	"time"
 
-	"github.com/rs/zerolog/log"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
-	"github.com/stretchr/testify/require"
-	"go.uber.org/fx"
-
 	"github.com/dadrus/heimdall/internal/config"
-	"github.com/dadrus/heimdall/internal/heimdall"
 	"github.com/dadrus/heimdall/internal/rules/event"
 	"github.com/dadrus/heimdall/internal/x"
+	"github.com/rs/zerolog/log"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-type mockLifecycle struct{ mock.Mock }
-
-func (m *mockLifecycle) Append(hook fx.Hook) { m.Called(hook) }
-
-func TestRegisterFileSystemProvider(t *testing.T) {
-	t.Parallel()
-
-	tmpFile, err := os.CreateTemp(os.TempDir(), "test-rule-")
-	require.NoError(t, err)
-
-	defer os.Remove(tmpFile.Name())
-
-	for _, tc := range []struct {
-		uc         string
-		conf       config.Configuration
-		setupMocks func(t *testing.T, mockLC *mockLifecycle)
-		assert     func(t *testing.T, err error)
-	}{
-		{
-			uc: "without it being configured",
-			assert: func(t *testing.T, err error) {
-				t.Helper()
-
-				assert.NoError(t, err)
-			},
-		},
-		{
-			uc: "without provided rules file/directory",
-			conf: config.Configuration{
-				Rules: config.RulesConfig{
-					Providers: config.RuleProviders{
-						FileSystem: &config.FileBasedRuleProviderConfig{},
-					},
-				},
-			},
-			assert: func(t *testing.T, err error) {
-				t.Helper()
-
-				require.Error(t, err)
-				assert.ErrorIs(t, err, heimdall.ErrConfiguration)
-			},
-		},
-		{
-			uc: "with not existing referenced file",
-			conf: config.Configuration{
-				Rules: config.RulesConfig{
-					Providers: config.RuleProviders{
-						FileSystem: &config.FileBasedRuleProviderConfig{Src: "foo.bar"},
-					},
-				},
-			},
-			assert: func(t *testing.T, err error) {
-				t.Helper()
-
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), "no such file")
-			},
-		},
-		{
-			uc: "with existing rules file",
-			conf: config.Configuration{
-				Rules: config.RulesConfig{
-					Providers: config.RuleProviders{
-						FileSystem: &config.FileBasedRuleProviderConfig{Src: tmpFile.Name()},
-					},
-				},
-			},
-			setupMocks: func(t *testing.T, mockLC *mockLifecycle) {
-				t.Helper()
-
-				mockLC.On("Append", mock.AnythingOfType("fx.Hook"))
-			},
-			assert: func(t *testing.T, err error) {
-				t.Helper()
-
-				require.NoError(t, err)
-			},
-		},
-		{
-			uc: "with existing rules file and enabled watcher",
-			conf: config.Configuration{
-				Rules: config.RulesConfig{
-					Providers: config.RuleProviders{
-						FileSystem: &config.FileBasedRuleProviderConfig{Src: tmpFile.Name(), Watch: true},
-					},
-				},
-			},
-			setupMocks: func(t *testing.T, mockLC *mockLifecycle) {
-				t.Helper()
-
-				mockLC.On("Append", mock.AnythingOfType("fx.Hook"))
-			},
-			assert: func(t *testing.T, err error) {
-				t.Helper()
-
-				require.NoError(t, err)
-			},
-		},
-	} {
-		t.Run("case="+tc.uc, func(t *testing.T) {
-			// GIVEN
-			mlc := &mockLifecycle{}
-			queue := make(event.RuleSetChangedEventQueue, 10)
-			setupMocks := x.IfThenElse(tc.setupMocks != nil,
-				tc.setupMocks,
-				func(t *testing.T, mockLC *mockLifecycle) { t.Helper() })
-
-			setupMocks(t, mlc)
-
-			// WHEN
-			err := registerProvider(registrationArguments{Lifecycle: mlc, Config: tc.conf, Queue: queue}, log.Logger)
-
-			// THEN
-			tc.assert(t, err)
-
-			mlc.AssertExpectations(t)
-		})
-	}
-}
-
 // nolint: maintidx
-func TestStartFileSystemProvider(t *testing.T) {
+func TestStartProvider(t *testing.T) {
 	t.Parallel()
 
 	var tearDownFuncs []func()
@@ -151,21 +27,21 @@ func TestStartFileSystemProvider(t *testing.T) {
 
 	for _, tc := range []struct {
 		uc             string
-		createProvider func(t *testing.T, file *os.File, dir string) *fileSystemProvider
+		createProvider func(t *testing.T, file *os.File, dir string) *provider
 		writeContents  func(t *testing.T, file *os.File, dir string)
-		assert         func(t *testing.T, err error, provider *fileSystemProvider)
+		assert         func(t *testing.T, err error, provider *provider)
 	}{
 		{
 			uc: "start provider using not existing file",
-			createProvider: func(t *testing.T, file *os.File, dir string) *fileSystemProvider {
+			createProvider: func(t *testing.T, file *os.File, dir string) *provider {
 				t.Helper()
 
-				return &fileSystemProvider{
+				return &provider{
 					src:    "foo.bar",
 					logger: log.Logger,
 				}
 			},
-			assert: func(t *testing.T, err error, provider *fileSystemProvider) {
+			assert: func(t *testing.T, err error, provider *provider) {
 				t.Helper()
 
 				require.Error(t, err)
@@ -174,17 +50,17 @@ func TestStartFileSystemProvider(t *testing.T) {
 		},
 		{
 			uc: "start provider using file without read permissions",
-			createProvider: func(t *testing.T, file *os.File, dir string) *fileSystemProvider {
+			createProvider: func(t *testing.T, file *os.File, dir string) *provider {
 				t.Helper()
 
 				require.NoError(t, file.Chmod(0o200))
 
-				return &fileSystemProvider{
+				return &provider{
 					src:    file.Name(),
 					logger: log.Logger,
 				}
 			},
-			assert: func(t *testing.T, err error, provider *fileSystemProvider) {
+			assert: func(t *testing.T, err error, provider *provider) {
 				t.Helper()
 
 				require.Error(t, err)
@@ -193,16 +69,16 @@ func TestStartFileSystemProvider(t *testing.T) {
 		},
 		{
 			uc: "successfully start provider without watcher using empty file",
-			createProvider: func(t *testing.T, file *os.File, dir string) *fileSystemProvider {
+			createProvider: func(t *testing.T, file *os.File, dir string) *provider {
 				t.Helper()
 
-				return &fileSystemProvider{
+				return &provider{
 					src:    file.Name(),
 					logger: log.Logger,
 					queue:  make(event.RuleSetChangedEventQueue, 10),
 				}
 			},
-			assert: func(t *testing.T, err error, provider *fileSystemProvider) {
+			assert: func(t *testing.T, err error, provider *provider) {
 				t.Helper()
 
 				require.NoError(t, err)
@@ -212,19 +88,19 @@ func TestStartFileSystemProvider(t *testing.T) {
 		},
 		{
 			uc: "successfully start provider without watcher using not empty file",
-			createProvider: func(t *testing.T, file *os.File, dir string) *fileSystemProvider {
+			createProvider: func(t *testing.T, file *os.File, dir string) *provider {
 				t.Helper()
 
 				_, err := file.Write([]byte(`Hi Bar`))
 				require.NoError(t, err)
 
-				return &fileSystemProvider{
+				return &provider{
 					src:    file.Name(),
 					logger: log.Logger,
 					queue:  make(event.RuleSetChangedEventQueue, 10),
 				}
 			},
-			assert: func(t *testing.T, err error, provider *fileSystemProvider) {
+			assert: func(t *testing.T, err error, provider *provider) {
 				t.Helper()
 
 				require.NoError(t, err)
@@ -240,16 +116,16 @@ func TestStartFileSystemProvider(t *testing.T) {
 		},
 		{
 			uc: "successfully start provider without watcher using empty dir",
-			createProvider: func(t *testing.T, file *os.File, dir string) *fileSystemProvider {
+			createProvider: func(t *testing.T, file *os.File, dir string) *provider {
 				t.Helper()
 
-				return &fileSystemProvider{
+				return &provider{
 					src:    dir,
 					logger: log.Logger,
 					queue:  make(event.RuleSetChangedEventQueue, 10),
 				}
 			},
-			assert: func(t *testing.T, err error, provider *fileSystemProvider) {
+			assert: func(t *testing.T, err error, provider *provider) {
 				t.Helper()
 
 				require.NoError(t, err)
@@ -259,7 +135,7 @@ func TestStartFileSystemProvider(t *testing.T) {
 		},
 		{
 			uc: "successfully start provider without watcher using dir with not empty file",
-			createProvider: func(t *testing.T, file *os.File, dir string) *fileSystemProvider {
+			createProvider: func(t *testing.T, file *os.File, dir string) *provider {
 				t.Helper()
 
 				tmpFile, err := os.CreateTemp(dir, "test-rule-")
@@ -268,13 +144,13 @@ func TestStartFileSystemProvider(t *testing.T) {
 				_, err = tmpFile.Write([]byte(`Hi Foo`))
 				require.NoError(t, err)
 
-				return &fileSystemProvider{
+				return &provider{
 					src:    dir,
 					logger: log.Logger,
 					queue:  make(event.RuleSetChangedEventQueue, 10),
 				}
 			},
-			assert: func(t *testing.T, err error, provider *fileSystemProvider) {
+			assert: func(t *testing.T, err error, provider *provider) {
 				t.Helper()
 
 				require.NoError(t, err)
@@ -290,7 +166,7 @@ func TestStartFileSystemProvider(t *testing.T) {
 		},
 		{
 			uc: "successfully start provider without watcher using dir with other directory with rule file",
-			createProvider: func(t *testing.T, file *os.File, dir string) *fileSystemProvider {
+			createProvider: func(t *testing.T, file *os.File, dir string) *provider {
 				t.Helper()
 
 				tmpDir, err := os.MkdirTemp(dir, "test-dir-")
@@ -302,13 +178,13 @@ func TestStartFileSystemProvider(t *testing.T) {
 				_, err = tmpFile.Write([]byte(`Hi Foo`))
 				require.NoError(t, err)
 
-				return &fileSystemProvider{
+				return &provider{
 					src:    dir,
 					logger: log.Logger,
 					queue:  make(event.RuleSetChangedEventQueue, 10),
 				}
 			},
-			assert: func(t *testing.T, err error, provider *fileSystemProvider) {
+			assert: func(t *testing.T, err error, provider *provider) {
 				t.Helper()
 
 				require.NoError(t, err)
@@ -319,7 +195,7 @@ func TestStartFileSystemProvider(t *testing.T) {
 		{
 			uc: "successfully start provider with watcher using initially empty dir and adding rule " +
 				"file and deleting it then",
-			createProvider: func(t *testing.T, file *os.File, dir string) *fileSystemProvider {
+			createProvider: func(t *testing.T, file *os.File, dir string) *provider {
 				t.Helper()
 
 				provider, err := newProvider(
@@ -348,7 +224,7 @@ func TestStartFileSystemProvider(t *testing.T) {
 
 				time.Sleep(200 * time.Millisecond)
 			},
-			assert: func(t *testing.T, err error, provider *fileSystemProvider) {
+			assert: func(t *testing.T, err error, provider *provider) {
 				t.Helper()
 
 				require.NoError(t, err)
@@ -374,7 +250,7 @@ func TestStartFileSystemProvider(t *testing.T) {
 		{
 			uc: "successfully start provider with watcher using initially empty file, " +
 				"updating it afterwards and deleting it then",
-			createProvider: func(t *testing.T, file *os.File, dir string) *fileSystemProvider {
+			createProvider: func(t *testing.T, file *os.File, dir string) *provider {
 				t.Helper()
 
 				provider, err := newProvider(
@@ -398,7 +274,7 @@ func TestStartFileSystemProvider(t *testing.T) {
 
 				time.Sleep(200 * time.Millisecond)
 			},
-			assert: func(t *testing.T, err error, provider *fileSystemProvider) {
+			assert: func(t *testing.T, err error, provider *provider) {
 				t.Helper()
 
 				require.NoError(t, err)
